revert long variable conversion

+++ core/src/main/java/site/ycsb/ClientThread.java
@@ -34,10 +34,10 @@ public class ClientThread implements Runnable {
   private DB db;
   private boolean dotransactions;
   private Workload workload;
-  private long opcount;
+  private int opcount;
   private double targetOpsPerMs;
 
-  private long opsdone;
+  private int opsdone;
   private int threadid;
   private int threadcount;
   private Object workloadstate;
@@ -56,7 +56,7 @@ public class ClientThread implements Runnable {
    * @param targetperthreadperms target number of operations per thread per ms
    * @param completeLatch        The latch tracking the completion of all clients.
    */
-  public ClientThread(DB db, boolean dotransactions, Workload workload, Properties props, long opcount,
+  public ClientThread(DB db, boolean dotransactions, Workload workload, Properties props, int opcount,
                       double targetperthreadperms, CountDownLatch completeLatch) {
     this.db = db;
     this.dotransactions = dotransactions;
@@ -81,7 +81,7 @@ public class ClientThread implements Runnable {
     threadcount = threadCount;
   }
 
-  public long getOpsDone() {
+  public int getOpsDone() {
     return opsdone;
   }
 
@@ -179,8 +179,8 @@ public class ClientThread implements Runnable {
   /**
    * The total amount of work this thread is still expected to do.
    */
-  long getOpsTodo() {
-    long todo = opcount - opsdone;
+  int getOpsTodo() {
+    int todo = opcount - opsdone;
     return todo < 0 ? 0 : todo;
   }
 }
